extract more common
